refactor: move ecstore rebalance runtime loop
